requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
pydantic[email]
python-multipart
